New Operation data function
new override function in operation data for data in load

+++ Assets/Data/Scriptables Model/UserData.cs
@@ -60,7 +60,7 @@ public class UserData : ScriptableObject
 
     [Header("----Punto 1")]
     public int unitsProccess;
-    public float requiredUnits;
+    public int requiredUnits;
     [Header("----Punto 2")]
     public int kSuplements;
